refactor: 엔티티 Integer -> Long 수정, 해당 변경에 따른 코드 변경, fix: email 인증 숫자 생성 시 6자리 숫자 대신 5자리로 출력되는 문제 해결

+++ src/main/java/com/travelservice/domain/auth/entity/PasswordResetRequest.java
@@ -27,7 +27,7 @@ import lombok.NoArgsConstructor;
 public class PasswordResetRequest {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer requestId;
+	private Long requestId;
 
 	@Column(nullable = false, length = 100)
 	private String email;
